Fix #6 - code refactoring
웹소켓의 on.Message에 있던 activity 파싱 코드를 utils 하부의 processActivies 로 분리하였습니다.
